Move course enrollment related methods to service
