+++ app/views/studentship/course_enrollments/_student_info.html.erb
@@ -18,15 +18,15 @@
       </tr>
       <tr>
         <th><%= t('.academic_term') %></th>
-        <td><%= full_name(@student.active_term) %></td>
+        <td><%= full_name(@service.active_term) %></td>
       </tr>
       <tr>
         <th><%= t('.gpa') %></th>
-        <td><%= @student.fake_gpa %></td>
+        <td><%= @student.gpa %></td>
       </tr>
       <tr>
         <th><%= t('.enrollment_status') %></th>
-        <td><%= t('activerecord.enums.course_enrollment.statuses')[@student.enrollment_status] %></td>
+        <td><%= t('activerecord.enums.course_enrollment.statuses')[@service.enrollment_status] %></td>
       </tr>
     </tbody>
   </table>
