+++ app/views/studentship/course_enrollments/index.html.erb
@@ -10,7 +10,7 @@
       </div>
       <div class="card-body">
         <div class="list-group">
-          <% if @student.enrollment_status != :saved %>
+          <% if @service.enrollment_status != :saved %>
             <% url = new_student_course_enrollment_url(@student) %>
           <% else %>
             <% url = list_student_course_enrollments_url(@student) %>
@@ -21,7 +21,7 @@
               <h5 class="mb-1"><%= @student.unit.name %></h5>
             </div>
             <p class="mb-1"><%= @student.registation_date_range %></p>
-            <small><%= "#{t('.enrollment_status')}: #{t('activerecord.enums.course_enrollment.statuses')[@student.enrollment_status]}" if @student.enrollment_status.presence %></small>
+            <small><%= "#{t('.enrollment_status')}: #{t('activerecord.enums.course_enrollment.statuses')[@service.enrollment_status]}" if @service.enrollment_status.presence %></small>
           <% end %>
         </div>
       </div>
