+++ test/decorators/student_decorator_test.rb
@@ -14,20 +14,13 @@ class StudentDecoratorTest < ActiveSupport::TestCase
 
   test 'registation_date_range method' do
     calendar = calendars(:bm_calendar)
+    event = calendar_events(:active_online_course_registrations)
     assert_equal StudentDecorator.new(@student).registation_date_range,
-                 translate('index.registration_date_range', calendar.date_range('online_course_registrations'))
-  end
-
-  test 'enrollment_status method' do
-    assert_not StudentDecorator.new(students(:serhat_omu)).enrollment_status
-    assert_equal StudentDecorator.new(students(:john)).enrollment_status, :saved
-    assert_equal @student.enrollment_status, :draft
-  end
-
-  test 'course_catalog method' do
-    courses = @student.course_catalog.first[:courses]
-    assert_includes courses, available_courses(:elective_course_2)
-    assert_includes courses, available_courses(:compulsory_course_2)
+                 translate(
+                  'index.registration_date_range',
+                  start_time: event.start_time&.strftime('%F %R'),
+                  end_time:   event.end_time&.strftime('%F %R')
+                )
   end
 
   private
